app/requirements.txt
pyedifice>=5.0.0
PySide6>=6.5
requests>=2.28
openpyxl>=3.1
